app/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
python-multipart==0.0.6
numpy<2